feat: Add image click handler to Carousel component
The Carousel component now includes a click handler for the images. When an image is clicked, the active state is updated to reflect the clicked image. This allows users to navigate through the carousel by clicking on the images.

This commit also introduces a new Modal component and adds a "Adopt" button to the Details component. Clicking on the "Adopt" button opens a modal dialog asking the user if they would like to adopt the pet. The modal dialog includes "Yes" and "No" buttons for the user to make a selection.

Lastly, the index.html file is modified to include a div element for rendering the modal component.